Add files via upload
the contents herein are to be used solely at your own risk. this is fully a beta test version only, please use safe practices when storing security keys. when using the snap.js do not store it in the front end app, bare minimum store outside of the app if not run it on a separate device entirely. if you use the contents herein please note it is highly advided that you edit the snap.js to call the security key from a safe location in a safe manner, this is only a test version.
 YOU MUST CREATE A SECONDARY ADMIN ACCOUNT WHEN DEPLOYING CONTRACT KEEP ADMIN KEYS SECURE